requirements.txt
# absl-py
# altair
# astunparse
# attrs
# blinker
# cachetools
# certifi
# charset-normalizer
# click
# colorama
# contourpy
# cycler
# flatbuffers
# fonttools
# gast
# gitdb
# GitPython
# google-auth
# google-auth-oauthlib
# google-pasta
# grpcio
# h5py
# idna
# importlib-metadata
# jax
# Jinja2
# joblib
# jsonschema
# jsonschema-specifications
# keras
# kiwisolver
# libclang
# Markdown
# markdown-it-py
# MarkupSafe
# matplotlib
# mdurl
# ml-dtypes
# nltk
# numpy
# oauthlib
# opt-einsum
# packaging
# pandas
# Pillow
# protobuf
# pyarrow
# pyasn1
# pyasn1-modules
# pydeck
# Pygments
# Pympler
# pyparsing
# python-dateutil
# pytz
# pytz-deprecation-shim
# referencing
# regex
# requests
# requests-oauthlib
# rich
# rpds-py
# rsa
# scipy
# six
# smmap
# streamlit
# tenacity
# tensorboard
# tensorboard-data-server
# tensorflow
# tensorflow-estimator
# tensorflow-io-gcs-filesystem
# termcolor
# toml
# toolz
# tornado
# tqdm
# typing_extensions
# tzdata
# tzlocal
# urllib3
# validators
# watchdog
# Werkzeug
# wrapt
# zipp
pip
streamlit==0.88.0
Pillow==8.4.0
tensorflow==2.7.0
numpy==1.19.5
keras
